Move background selection to its own file.
